Move fixtures into testing dir
